feat: Remove core module and separate handler from feature
