+++ personalization-sdk/src/main/kotlin/com/personalization/features/notification/presentation/helpers/NotificationHelper.kt
@@ -5,9 +5,9 @@ import android.content.Context
 import android.graphics.Bitmap
 import android.widget.RemoteViews
 import androidx.core.app.NotificationCompat
-import com.personalization.features.notification.core.NotificationResources
 import com.personalization.features.notification.domain.model.NotificationConstants.NOTIFICATION_CHANNEL
 import com.personalization.features.notification.domain.model.NotificationData
+import com.personalization.resources.NotificationResources
 
 object NotificationHelper {
